deleted default generated files and fixed some syntax/formatting
