+++ foodie-spot/src/components/types.ts
@@ -1,0 +1,7 @@
+export interface Restaurant {
+  restaurantName: string;
+  foodType: string;
+  location: string;
+  hours: string;
+  rating: number;
+}
