+++ foodie-spot/src/pages/Home.tsx
@@ -1,12 +1,8 @@
-import React from 'react';
-
-const Home: React.FC = () => {
+export default function Home() {
   return (
     <div>
       <h1>Welcome to Foodie Spot</h1>
       <p>Find delicious meal plans and restaurants around you!</p>
     </div>
   );
-};
-
-export default Home;
+}
